requirements.txt
langchain
pypdf
tiktoken 
openai
faiss-cpu
flask
python-dotenv
gunicorn